+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/TestEngineRegistry.java
@@ -1,4 +1,4 @@
-package com.teamscale.testimpacted.junit;
+package com.teamscale.test_impacted.engine;
 
 import org.junit.platform.commons.util.ClassLoaderUtils;
 import org.junit.platform.engine.TestEngine;
