+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/executor/AvailableTests.java
@@ -1,9 +1,9 @@
-package com.teamscale.testimpacted.junit.executor;
+package com.teamscale.test_impacted.engine.executor;
 
 import com.teamscale.client.ClusteredTestDetails;
 import com.teamscale.client.TestDetails;
 import com.teamscale.client.TestForPrioritization;
-import com.teamscale.testimpacted.junit.ImpactedTestEngine;
+import com.teamscale.test_impacted.engine.ImpactedTestEngine;
 import org.conqat.lib.commons.string.StringUtils;
 import org.junit.platform.commons.logging.Logger;
 import org.junit.platform.commons.logging.LoggerFactory;
