+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/options/ServerOptions.java
@@ -1,4 +1,4 @@
-package com.teamscale.testimpacted.junit.options;
+package com.teamscale.test_impacted.engine.options;
 
 import org.junit.platform.commons.util.Preconditions;
