+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/test_descriptor/JUnitVintageTestDescriptorResolver.java
@@ -1,4 +1,4 @@
-package com.teamscale.testimpacted.test_descriptor;
+package com.teamscale.test_impacted.test_descriptor;
 
 import org.junit.platform.engine.TestDescriptor;
 import org.junit.platform.engine.TestEngine;
