+++ impacted-test-engine/src/main/resources/META-INF/services/org.junit.platform.engine.TestEngine
@@ -1,1 +1,1 @@
-com.teamscale.testimpacted.junit.ImpactedTestEngine
+com.teamscale.test_impacted.engine.ImpactedTestEngine
